Simplify `clang_getEnumConstantDeclValue`
In order to do the error check we need to check the cursor kind; previously we
were (incorrectly?) looking at the _type_ kind, which for an enum member will
just report `Int`, which isn't particularly informative. We therefore looked at
the type kind of the _parent_, but instead we should just look at the kind of
the _cursor_ (which will be `EnumConstantDecl`).