Start work on prelude and parsing macros
The idea is to have some bootstrapping code as part of `hs-bindgen` that trawls
the C standard headers looking for type definitions (such as `uint64_t`); these
types should become part of a prelude, which we can recognize out of the box
(so that instead of generating a new typedef for `uint64_t` every time we
generate bindings, we use the one and only from the prelude).

Since the standard headers also contain a whole bunch of macros, and we anyway
want to support generating bindings for specific forms of macros (#43), this
also has the beginnings of a parser for C macros.