Delete `foldRaw` and `foldComments`
Also delete the corresponding golden test (that was dumping the `Forest
Element`): depending on the `clang` version as well as the platform (Linux vs
OSX) this tree may look different, but the C AST that we construct looks the
same in all cases (even in the presence of standard platform types that are
sometimes macros and sometimes typedefs). This also cleans up the command line
a bit; all dev tools are now under `dev`.