Add clang-ast-dump package (#212)
This commit is a squashed version of a number of previous commits,
rebased to work with a recent redesign of `Fold` as well as the
reorganized module hierarchy.

`FoldM` uses `MonadIO`, and I refactored the code to call `liftIO`
minimizing changes.  Now that we can use a `Reader`, it is possible to
track indentation within the `Reader` environment, using `local` for
indentation, but I have not implemented that.

The new version of `Fold` no longer provides `parent`, as the parent
cursor can be queried using the API.  I rewrote the code to query both
the semantic and lexical parents.  When they are equal and are not the
target file, just "parent" is displayed.  Otherwise, "semantic parent"
and "lexical parent" are displayed separately.