Throw errors on unexpected comment kinds (#26)
We decided against lenient parsing of `libclang` comments and now throw
an error upon encountering an unexpected comment kind, such as block
content within inline content.

We will of course need to test this against various real-world code.
Error messages include the cursor display name, the file name, and the
extent.  We can investigate the `libclang` comment AST for any errors
using the `clang-ast-dump` utility.

Note that this internal module is not used anywhere yet, so it is
therefore untested.  I will test it when returning to the documentation
task.