Add type spelling to ASTs
This commit adds a type spelling field to the following ASTs:

* `C` phase: `Struct`, `Enu`, `Typedef`
    * Type: `Text`
* `Hs` phase: `Struct`, `Newtype`
    * Type: `Maybe Text`
* `SHs` phase: `Record`, `Newtype`
    * Type: `Maybe Text`

As suggested in #316, the string is not parsed.  This commit simply uses
`Text`, but we could implement a `newtype` wrapper if desired.

The goal is to make it easier to generate tests for structures,
enumerations, and `typedef`s of structures/enumerations.  Note that this
is also required for unions, but those are not implemented yet.  This is
the minimal change required to do this; the `Maybe` is needed because it
does *not* track the type spelling for macros.

(Cherry-picked from `source-info` for experimentation)