Implement C header and source generation (WIP)
This is an experiment using the C type spelling to "join" the translated
Haskell declarations with corresponding C sources.  Doing it this way
provides a lot of room for error.  This commit is an initial
implementation with limited support and lots of room for improvement.

Note that `Ord` instances are derived for `SHs.Global` and `HsPrimType`
in order to put them in a `Set`.