Implement minimal default name mangling (#331)
This commit changes name mangling so that there are minimal changes by
default, not attempting to create Haskell-style names.  One significant
difference with the previous implementation is that it can now add a
prefix if/when the first character is not valid, used to handle leading
underscores when creating type constructors.

`MkHsName` is removed, replaced by `mkHsNamePrefixInvalid` and
`mkHsNameDropInvalid` functions that are now passed to `translateName`
like other options.

`ctxFieldVarSingleConstr` is removed since constructors now have the
same name as types by default, and we do not generate sum types anyway.
If we do so in the future, we need to decide how to do name mangling for
constructors as well as accessors.

`HsBindgen.C.Fold.Type.mkDefnName` is changed to follow the conventions
of the new defaults.