Change name mangling overrides
The initial implementation allowed users to override name mangling by
specifying the C name.  This is problematic when the name of an
anonymous structure/union needs to be overridden.  Specifying a C name
worked when some name mangling was performed while creating the C AST,
but it does not work when all name mangling is done by the name mangler.
With this change, users override by specifying the translated Haskell
name.

Overrides must now be performed after creating the Haskell name.  The
parameter order of `translateName` is updated for consistency.