Refactor struct name mangling
This commit removes `DefnName` and instead tracks the "path" of a
structure, renaming the type to `DeclPath`.

`mkDefnName` is now implemented as part of name mangling (as
`getDeclPathParts`), in a way that works with name mangling options.

Context `AnonNamedFieldTypeConstrContext` is removed, replaced with
`StructTypeConstrContext` that contains a `DeclPath` instead of a
`CName`.  New function `translateDeclPath` is a high-level function
for translating `DeclPath`s.