Add DeclName
When generating C code, we have to know how to write each C name (such
as `struct foo` vs. just `foo`).  This commit adds type `DeclName` for
use in the `DeclPathStruct` constructor of type `DeclPath` to do this.

Note that this could alternatively be done by making `CName` a phantom
type that specifies the namespace.  I implemented `DeclName` for now as
a simple/minimal change.