Render Hs AST using Data.TreeDiff.Expr
The motivation for this is to normalize representations just for tests.
`SourcePath` values are rendered as lists, like in the `tree-diff`
fixtures.

Note that `ediffGolden1` is *not* used because we do not have `Eq`
instances for the `Hs` AST.  We could use it if we decide to implement
`Eq` instances.  For now, per-line string comparisons are made.  This is
still an improvement over the previous implementation using `Show`
because the `Expr`s are pretty-printed across multiple lines.