Parse bitfields
Do not yet generate any Haskell code for them though.
We store bitfields separately from rest of the fields.
An idea is that in Haskell representation we could have a single
field in the data-record for all the bitfields,
even if they are spread across / interleaved between other fields
in the C struct.