Implement relative paths in constructors
We decided to implement relative path support when constructing a
`SourcePath`, so that it is consistent across the whole application.

The directory to make paths relative to is passed as configuration
throughout the whole application, so that we can enable users to
configure it.  For now, I implemented it by consistently passing it as
the first parameter to functions (`Reader`-style).  Every parameter is
named `relPath` and is documented in the same way.  If we implement more
configuration the future, we may want to consider using a `Reader`
transformer.

Threading configuration through an application changes the API for many
functions. orz