Rewrite scripts/musl-update.sh
The headers can be installed *without* any compilation if configuration
is done manually, not using the `configure` script since it checks for
cross-compilation toolchains.

The script no longer has to run on an `x86_64` architecture.  Since
toolchains are no longer needed, this enables us to easily vendor
headers for more architectures.