Add dev prelude --include option
Setting an include directory is required for the command to work.  This
option sets a default when there are vendored Musl headers for the host
architecture.  When there are not vendored Musl headers for the host
architecture, the user must specify the option.

Option `-nostdinc` is now always set for this command.  Oddly, this
option is not displayed in the `--help` output on my system, but there
is no error when using it, and I see it in the current documentation
online.

The include directory is passed to Clang via `-isystem`.