Fix function argument iteration
Problematic code:

```haskell
nargs <- liftIO $ clang_getNumArgTypes ty
args' <- forM [0 .. fromIntegral nargs - 1] $ \i -> ...
```

Function `clang_getNumArgTypes` returns an `int` while
`clang_getArgType` takes an `unsigned int`, hence the need for
`fromIntegral`.  This does not work when the function has no arguments,
however.  When `nargs` is zero, `fromIntegral nargs - 1` underflows to
the maximum value of the unsigned type.  This results in many
iterations, but it generally fails quickly when attempting to get and
use a non-existent argument.

This commit fixes the issue by simply moving the `fromIntegral` inside
of the iteration.

A function with no arguments is added to the `simple_func.h` example.