Deal with arrays of unknown size as funargs
This takes advantage of the fact that we now have context for the `C.Type ->
HsType` translation. It's also necessary to fix this before fixing the
translation of `PrimVoid`, because we were previously inferring `PrimVoid` as a
placeholder for these arrays, which would then result in a use of `void` in a
context in which we don't expect it, resulting in an exception.