Always use C.Expr.BuildPlatform
From the perspective of the compiler, it's always compiling code for
host. Thus we can use `C.Expr.BuildPlatform` (which is really a
HostPlatform - compiled code shouldn't really be aware of build platform
anwyay).

This is an alternative to #391. Arguably a lot simpiler one.
There are a lot of further cleanups, which can be made in `c-expr`,
(e.g. renaming the `C.Expr.BuildPlatform`, potentially getting rid of
non-host architecture modules, simplifying library structure, making
instances non-orphans etc) but these are not strictly necessary,
so not done to not obscure this change.