Add clang_getIncludedFile
I went ahead and implemented this using `hs-bindgen-bootstrap`, which
required moving the `CXFile` type into a submodule that can be imported
from both `Core` and `FFI`.  Instead of a single `Types` module, there
are separate modules for each type of type.  `CXFile` is a void pointer,
not an enumeration or structure, so I added a new `Pointers` module.