Move/refactor Ref types, removing Ref module
The C name spelling type is moved to `HsBindgen.Clang.CNameSpelling`.
The Aeson instance is moved into `HsBindgen.Orphans`.

C include search path directory and C header types and functions are
moved to `HsBindgen.Clang.Paths`.  A new `CIncludeAbsPathDir` type is
added.  Type `CIncludePathDir` may now be absolute or relative to the
current working directory, and function `resolveCIncludeAbsPathDirs` is
used to resolve to `CIncludeAbsPathDir`.  Allowing users to specify
include search path directories using relative paths is more
user-friendly, and it simplifies our preprocessor test, which specifies
paths in a Literate Haskell file with a *pure* list of arguments.  The
Aeson instance for `CHeaderRelPath` is moved into `HsBindgen.Orphans`.

Haskell module, package, and identifier types are moved into
`HsBindgen.ExtBindings`.

The awkward `HsBindgen.Ref` module is removed.

Clang should be passed absolute paths to user-specified headers so that
it consistently reports absolute paths.  This is not done yet, so a
temporary hack resolves relative paths relative to the current working
directory when building the include graph.  The hack is annotated with a
`TODO`; I will remove the hack when absolute paths are passed to Clang.