Revise ExtBindings representation
The initial implementation included an optimization that assumed that C
names within an execution of `hs-bindgen` are unique.  That is not
always the case.  C names within a translation unit must be unique, but
we may deal with different types that have the C name defined in
different headers that are intended to be used separately.

With this change, C names in different headers can map to different
Haskell identifiers.  It is only a change to the internal
representation; there is no change to the JSON/YAML schema.