Add external bindings interface
An `--external-bindings` option is added to the CLI.  It may be used
more than once to specify multiple external bindings configurations.

An external bindings parameter of type `[FilePath]` is added to the TH
API.

Both YAML and JSON files are supported.

Multiple external bindings are merged into a single `Map`.  An error is
raised if different external bindings contain configuration for the same
C name spelling and header.

Up next is actually using the external bindings.  See the `parseCHeader`
function in `HsBindgen.Lib`.