Update clang-ast-dump
I updated this because we wanted to check that continuing and recursing
on a `CXCursor_InclusionDirective` does the same thing.  It does.

To check this, various things were updated.  `foldDecls` was changed to
use `IO` directly, so many `liftIO` calls could be simplified.  Some
dumped information was moved into helper functions, making it easier to
control what is dumped depending on the type of cursor.  A case for
`CXCursor_InclusionDirective` was added.  Include search path options
were added, and the type of the input file was changed to
`CHeaderRelPath`.  There are more things to update, but this was
sufficient to perform the needed check.