Resolve headers using libclang
This commit completely changes how we deal with header paths and C
include paths.

Support for parsing in-memory headers using `CXUnsavedFile` is added.  A
high-level wrapper function is added to scope string allocation.

Bindings for `clang_disposeTranslationUnit` and `clang_disposeIndex` are
added.  High-level wrapper functions that use `bracket` to free
resources is added.  The codebase is updated to do this, except for
`clang-tutorial`.  The `c-expr` tests needed to be refactored a bit so
that the `libclang` functions are called within scope.

Header resolution is done by parsing an in-memory header that just
includes the header that needs to be resolved.  The fold breaks after
the inclusion directive has been processed.