Refactor test-th
This change prefixes the top-level definitions for tests, so that we do
not have to worry about conflicts between different tests.

Note that we could of course put different tests in different modules,
but we currently symlink the test file and would have to symlink all of
those modules.  Perhaps it would be preferable to put the tests in the
same source directory, so that we can use the same test code (without
symlinks).  The generated modules can be put in subdirectories, and the
Cabal configuration would use the appropriate one.