Add struct, union, and enum aliases
Aliases of type `[DeclPath]` are added to the C `Struct`, `Union`, and
`Enu` AST nodes.  These record C type spellings that reference the same
representation in our AST, a subset of the aliases registered as
`TypeDeclAlias` in the `DeclState`.

They are also added to `TypeDeclProcessing`.  This is needed to support
aliases when using forward references.

This commit adds `typedef union n {...} n;` support.  I will add a test
separately.