Make `GenerateName` a record
We have two concepts of "invalid", exemplified by `escapeInvalidChar` and
`mkHsNamePrefixInvalid`, referring to characters that are invalid /anywhere/ in
an identifier and characters that are invalid as the /first/ character,
respectively. I think the intend will become clearer if we combine these two
steps into one, as separate functions in `GenerateName`. This is the first step
in that direction.