Move implementation into hs-bindgen:internal
This commit implements the minimal change for moving the `hs-bindgen`
implementation into an internal library.  Currently, all modules are
exposed, except for those that are documented in the Cabal file as
being re-exported.  We should probably clean up the API *after* this
change has been merged and everybody has already rebased to resolve
any merge conflicts in other branches.

The public library implements the `Lib` and `TH` APIs as well as
re-exports the following modules:

* `HsBindgen.Errors`
* `HsBindgen.Hs.NameMangler`
* `HsBindgen.Hs.NameMangler.DSL`

If we do not want to expose all of the `Errors` module, we should
refactor that.

Note that I followed the suggestion to *not* use a unique module name
prefix (`HsBindgen.Internal`) for the internal library.  If issues
arise, I am happy to refactor the library to use one.

Executable `hs-bindgen-cli`, test `test-pp`, and test `test-th` do *not*
depend on the internal library.  I think that we should maintain this,
as these should only use/test the public API.

Development executable `hs-bindgen-dev`, development executable
`clang-ast-dump`, and test `golden` depend on the internal library, by
design.