Move clang-bootstrap into clang
There is a tight coupling, so it is nice to version them together.  This
also reduces the number of packages that will need to be maintained on
Hackage.

A `dev` flag is added, and executable `clang-bootstrap` is only built
when it is set.

Example workflow:

```
$ cabal build clang -f dev
$ cabal run clang-bootstrap -f dev
$ cabal build all -f dev
$ cabal test all
```

Just `cabal-bootstrap` can be built and run, but note that `cabal clean`
must then be run before further development:

```
$ cabal build clang:clang-bootstrap -f dev
$ cabal run clang-bootstrap -f dev
$ cabal clean
```