Replace `prefixDecon` with `partDecon`
As part of #527 it became apparent that we were somewhat inconsistent in the
default name mangler; given something like

```c
typedef int myint;
```

we would generate

```haskell
newtype Myint = Myint
  { unMyint :: FC.CInt
  }
```

but the default name mangler uses snake case, not camel case. This PR fixes
this, so that we now generate

```haskell
newtype Myint = Myint
  { un_Myint :: FC.CInt
  }
```

This makes the code very consistent (see the change to `ProduceCandidate`), and
I think the output also more consistent.