test ccall
It won't work as is, as in test-th (and test-pp) we use header-only
example library, so linker fails.

But this illustrates that we don't really need capi as such,
it's mostly an additional safety check.

I want to see that this fails the same way on different operating
systems, i.e. trigger CI.

work towards using userland-capi.