Deal with `CXType_Char_U`
When the user does not specify the sign for `char`, the compiler picks one
(this is different from `int`, where the absence of a sign means `signed`,
rather than "implementation defined"). We dealt with the CXType_Char_S` case,
which is where the inferred sign is `signed`, but didn't deal with the case
where the inferred sign is `CXType_Char_U`, or unsigned. Unfortunately I'm not
sure under which circumstances this happens so this commit is lacking a test
currently (but the client has one).