Use handwritten `FromJSON`/`ToJSON` instances
I think generic derivations of serialization code should be avoided, at least
for externally visible formats: it ties the external format to closely to the
internal representation. It also makes it more difficult to make specific
tweaks. In this commit we replace the derived `ToJSON`/`FromJSON` instances by
equivalent handwritten ones, in the next commit we'll tweak these instances
slightly.