Rename to avoid confusion with Storable
`HasStaticSize` is renamed to `StaticSize`, and the methods are renamed
to `staticSizeOf` and `staticAlignment`.

`Peekable` is renamed to `ReadRaw`, and the method is renamed to
`readRaw`.  Function `peekByteOff` is renamed to `readRawByteOff`,
function `peekElemOff` is renamed to `readRawElemOff`, and function
`maybePeek` is renamed to `maybeReadRaw`.

`Pokable` is renamed to `WriteRaw`, and the method is renamed to
`writeRaw`.  Function `pokeByteOff` is renamed to `writeRawByteOff`, and
function `pokeElemOff` is renamed to `writeRawElemOff`.

The instances for the `base` types are *already* defined in the module.
I had forgotten that.

Note that `ReadRawWithCtxt` is *not* implemented yet.