Add instances to external bindings
The `TypeClass` type is moved to `HsBindgen.ExtBindings`, renamed to
`HsTypeClass`.  Type classes `StaticSize`, `ReadRaw`, and `WriteRaw`
are added.  Instances (`Set HsTypeClass`) are added to the
`ExternalBinding` type.  The external binding configuration format is
updated to include this information in an `instances` field, which may
be omitted when empty.  Instances for the `base` and
`hs-bindgen-runtime` external bindings are defined.

When generating external bindings, instances are not yet implemented.
They will be implemented in a subsequent commit.