c-expr: prove instance availability
This commit modifies the c-expr package to add functions that prove
instance availability.
To illustrate, the c-expr library now exports a function of type:

  singSub :: SType ty1 -> SType ty2 -> (ty1 -> ty2 -> SubRes ty1 ty2)

which allows us to recover the implementation of subtraction from the
types. In essence, given the types, this allows us to look up the
instance.

Here SType is a newly added singleton type that matches up C types
with corresponding types in Foreign.C.Types. For example, we have:

  SArithmetic (SFloatLike SDoubleType) :: SType CDouble

This also fixes an oversight with the existing instances, as for e.g.
(&&) we were generating overlapping instances:

  instance Logical (Ptr ty1) (Ptr ty1)
  instance Logical (Ptr ty1) (Ptr ty2)

Now we only generate the latter, as it subsumes the former.

In the future, this functionality will allow writing an evaluator for
the macro language:

  - we first infer types with the existing macro typechecker,
  - we then use the inferred types to look up the instance method.