Update binding specifications design
The `Omittable` type is moved up.  Instead of wrapping one of the
properties of the value that is omitted, it now wraps the omitted value
itself.

The `HsRef` type (represented using `haskell` in the YAML) is removed,
and both the module and Haskell identifier are (separately) optional.
This allows users to customize none, one, either, or both of them.