Use suffix for BindingSpecs types
The `BindingSpecs` module is meant to be imported qualified.  We
initially decided to *not* use suffixes for types: we would write
`BindingSpecs.Type` instead of `BindingSpecs.TypeSpec`.  This turned
out to be a bit confusing, so we decided to go ahead and use a suffix.

* `TypeSpec`
* `InstanceSpec`
* `StrategySpec`
* `ConstraintSpec`

Note that `Spec` is singular.  A `TypeSpec` is a specification for a
single type, etc.  `BindingSpecs` is plural, as it is a container for
specifications for multiple bindings.