Refactor to use OmittedDecls
Previously, type `SourceMap p` mapped from `QualId p` to `SourcePath`
for all declarations, including those that are not selected by the
predicate.  This is refactored to type `OmittedDecls`, which is
essentially the same but is only used for declarations that are not
selected by the predicate.  The motivation is to reduce duplication.

The key of the `OmittedDecls` map is essentially `QualId RenameAnon`,
since we decided to only deal with top-level types, which cannot be
anonymous.  To avoid referencing the `RenameAnon` pass from within the
`Parse` pass, the equivalent type `(CName, Namespace)` is used instead.
Since `OmittedDecls` is not parameterized, the annotations do not need
to be renamed.

The `OmittedDecls` values are accumulated in the `Parse` pass.  This is
essentially a `Writer`, but the accumulator is included in the state for
simplicity and performance.

The `ResolveBindingSpec` pass checks selected and omitted declarations
separately.