Resolve binding specification in two passes
This was something that was lost during the translation to the
`ResolveBindingSpec` pass.  A reference to a declaration may come before
the declaration itself, via a pointer.  With this change, information
about external bindings and omitted types of declarations is accumulated
during a first pass, and then the types within each declaration is
resolved in a second pass.