Use include graph to get main header
The tracking of the current main header did not work.  All include
directives are processed first, which makes since because they are
macros that are expanded by the preprocessor, which runs first.  As a
result, we cannot track the current main header as we fold the Clang
AST.

We now determine the main header for a function using the include graph.
From the source path for the function, we do a depth-first traversal of
the include graph until we find a main header path.

This must be done for every function.  The implementation in
`Data.DynGraph.Labelled` is for arbitrary graphs and works with cycles,
but we could implement a more performant version that just works with
DAGs (which is the case for the include graph).  One simple optimization
is implemented: no search is performed when there is a single main
header.

With this change, we no longer need to process include directives in
`foldDecl`.  I moved the `skipBuiltin` back into `match` and changed
`evalPredicate` to return the match result, since we need to distinguish
the skip reasons so that we do not call `recordNonSelectedDecl` on a
builtin.  `foldDecl` is now even more simple!