Add origin and aliases
This commit just adds the origin and aliases to the AST.  Uses will
be implemented in later commits.

These are *not* used as `Id`, because `QualId` should not include
include an origin.

For declarations, both are added to `DeclInfo`.  This makes them easy to
use even though they are not needed for some kinds of declarations, such
as functions.  For types, the origin is added to `TypeStruct`,
`TypeUnion`, `TypeEnum`, and `TypeMacroTypedef` constructors.

`TypeMacroTypedef` needs an origin because it can be renamed.  A
corresponding constructor is added to the external AST, as suggested by
a TODO comment.  This made it possible to avoid adding an origin to
`TypeTypedef`, which does not need one.

We should look closer at the aliases for the `macro_typedef_scope.h`
golden test.