Distinguish anonymous types via tokens
Function `getDeclId` distinguishes *anonymous* and *named* declarations,
but the Clang meaning of *anonymous* is different from what we need.  We
consider a `struct`, `union`, or `enum` declaration *anonymous* if there
is no tag, even if there is a `typedef` for the type.

`clang_Cursor_isAnonymous` does not do what we need.  It returns `False`
for an anonymous declaration if there is a `typedef` for the type.

In older versions of LLVM, one could check `clang_getCursorSpelling` for
an empty result, but this has changed in later versions of LLVM.  It is
recommended to not parse cursor or type spellings for this purpose.

We instead check for non-tagged declarations.  Function `isNotTagged`
returns `True` if the cursor is for a `struct`, `union`, or `enum`
declaration and the underlying tokens starts with that keyword and a
left bracket.