Refactor C.Spelling
The goal of this change is to factor out the old textual `CSpelling`
representation and instead make use of the new `C.NameKind` type.  In
particular, `bindingSpecTypes` was defined as
`Map CSpelling [(Set header, ...)]`.  I experimented
with `Map CName [(C.NameKind, [(Set header, ...)])]`, but this is not
useful because user-configured aliases may have different names, so I
changed it to more straightforward type
`Map (CName, C.NameKind) [(Set header, ...)]`.  `(CName, C.NameKind)`
is used often, so I went ahead and defined it as `C.Spelling`.

Note that `C.Spelling` is the same as `QualId NameAnon`.  I will
experiment with refactoring this.

Previously, we did not parse the textual `CSpelling` representation, but
now we need to.  This added a new type of error, which made multiple
types of errors possible in many functions.  That was possible in one
place before, and I used a specific exception then.  That is no longer
sufficient with this change, and that is something that has been on my
list to improve anyway.  I added a `MultiException` type that makes it
easier to accumulate multiple errors.

Generated binding specifications are ordered differently because we the
type of the `Map` key is changed.  I confirmed that this accounts for
all of the changes to the fixtures.