Refactor QualId to QualDeclId and QualName
The parametricity of `QualId p` caused more problems than it solved.
Two two uses are distinct, so this change separates them.

Type `Id p ~ DeclId => QualId p` is used in the beginning passes of
the frontend.  New type `QualDeclId` replaces this usage.

Type `Id p ~ CName => QualId p` is the same as `C.Spelling`, which is
renamed to `C.QualName`.

This change improves consistency and IMHO makes the code easier to
understand.