Refactor `BindingSpec` to use traces
This commit implemented many things that were then ported to a different
PR that was merged first, and then this commit was rebased.

This commit originally added a new `TraceBindingSpec`.  When ported,
specific message type `ResolveBindingSpecMsg` was added as
`TraceResolveBindingSpec`.  This rebased commit fixes this to use
`TraceBindingSpec` with the `BindingSpecMsg` sum type.

The `Exception` types are renamed to indicate that they are trace
messages.  (Example: `ReadBindingSpecMsg`)  They are currently all
errors, but we will likely add other types of traces in the future.

The (upstream) YAML error types do not have `Eq` instances, but we now
need `Eq` instances for the `Trace` type.  I resolved this by rendering
those errors and just storing `String`s.

The functions are now even more lenient than before.  For example, an
empty binding specification is returned when unable to load a file at
all.

Type `MultiException` is removed, as we are now managing errors via
traces.

This change means that we need tracers in more places.  Notable, the
tracer is in `Opts` but not `PPOpts`.  We have discussed making `Opts`
opaque and instead using new `Config` types for configuration.  Perhaps
`Opts` and `PPOpts` will be combined, resolving this issue.  For now, I
simply pass both where `PPOpts` was previously sufficient.