Refactor resolve tracing/errors
This commit implemented many things that were then ported to a different
PR that was merged first, and then this commit was rebased.

The original change was motivated by two things:

* We would like to trace header resolution, both successes and failures.
* Failure to resolve is an error in some cases and a warning in others.
  (See #777)

Type `ResolveHeaderException` was removed, and the API now returns a
`Maybe`, where `Nothing` indicates failure.  The caller can decide how
to handle failure.

Type `ResolveHeaderMsg` was added.  Both successful resolution and
failures are traced.