Resolve #279. Cleanup DVar treatment.
Treat DVar as "a binding (with type)".
I opened #279 because I didn't like that it was the only statement,
which required two TH declarations when translated.
It's not true anymore, as pattern synonyms also require two TH
declarations (similarly, type and "term").

I think it's fine to have it this way, and by *always* requring
the type we avoid conditional treatment.
If we would ever need top-level bindings ("vars") without
type annotations (which I think we shouldn't need), then DVar
could be split into type-signature and declaration.