Add support for global variables
This also adds plumbing for constants in the front-end, but does not yet
deal with them in the backend.

As I was working on this, I also realized that we resolved #684 in a
slightly suboptimal way: in principle we _can_ support _named_
declarations inside function signatures, and WG14-N3037 suggests that
perhaps we should. We therefore leave it up to the C compiler to warn
for this case or not; either way, we can generate code. One benefit of
fixing this is that the parser for global variables and the parser for
function signatures is now very very similar.

Other fixes in this PR:

* `requiredExtensions` was ignoring the type of foreign imports (important for
  example when those foreign imports use `ConstantArray`, and therefore needs
  `DataKinds`).
* Golden test infrastructure was not showing any messages on an exception;
  this makes it more difficult to see what's happening when `rust-bindgen`
  fails.

Closes #42