Improve definition of `UseOfDecl`
The new definition rules out some impossible definitions; we still have a panic
when these impossible definitions occur, but it would now arise when we
_construct_ these values, rather than where we consume them; if we are wrong
about them being impossible, we can provide a more useful error message there.

With the new definition it is easier to add support for other declarations (such as globals).