Add public opaque BindingSpec type
As discussed, new type `BindingSpec` wraps *both*
`UnresolvedBindingSpec` and `ResolvedBindingSpec`, so that it can be
used for the `stdlib` binding specification as well.  The public YAML
hack has been removed.

The `Pipeline.Opts` is public and configures binding specifications, so
this new type must also be defined within `hs-bindgen:internal`.  This
commit defines it in the `Pipeline` module.  We will refactor this a bit
when we make `Pipeline.Opts` internal, and this constraint will no
longer be an issue if we make the `Config` type specify filenames and
load/merge/resolve binding specifications internally.