Fix #795
This fix only works for `clang >= 19.1.0`, due to a bug in earlier versions
dealing with locations (https://github.com/llvm/llvm-project/pull/72400).
See `Clang.HighLevel.UserProvided.clang_getCursorSpelling` for a detailed
discussion of the fix.

In order to properly detect the `clang` version (and be able to distinguish
between clang 19.1 and before), I needed to change how we get the `clang`
version number. Since we're now parsing something that is explicitly described
as "should not be parsed", I've added some tests and some plausibility checks
(unfortunately, `clang` says that the version string should not be parsed but
does not provide an feasible alternative).