Simplify evaluation of a selection predicate
We were trying to construct a _reason_ why a selection predicate matched or
failed, but I realized that this is rather difficult to do well in the presence
of arbitrary boolean expressions, and indeed we were not doing it correctly
(though this was rather hidden: for example, logical conjunction was
implemented using monadic `>>`, effectively ignoring the reason why the first
predicate matched). In this PR we simplify this, mostly in preparation for
further changes.