Import HsBindgen.Language.C qualified as C
This module is for *standard* C language types, not types that are
specific to `hs-bindgen`.  It must always be imported qualified, and
sub-modules should be imported via this module, not directly.

Note that we also import other modules as qualified `C` as well,
sometimes within the same module.  We decided that this is fine.

This implements the long-pending change of renaming `CName` so that it
is used qualified as `C.Name`.