Add NsPrelimDeclId, refactor QualPrelimDeclId
The old version of `QualPrelimDeclId` was a product of `PrelimDeclId`
and `C.NameKind`.  This was not strictly correct because only the named
identities need the kind.  We do not need to track the kind of anonymous
identities, and we can assume that all builtin identities are ordinary.
The new version of `QualPrelimDeclId` associates the kind only with the
named variant.

Most uses of `QualPrelimDeclId` did not need to distinguish tag kinds,
so a new `NsPrelimDeclId` is added, which associates a `C.TypeNamespace`
with only the named variant.