Add DeclId
Type `PrelimDeclId` is used in the following passes.  The sum type
implicitly encodes origin information.

1. `Parse`
2. `Sort`
3. `Slice`
4. `HandleMacros`

Type `DeclId` is used in the following passes.  It is a product type
that includes a name and the origin.

5. `NameAnon`
6. `ResolveBindingSpec`
7. `HandleTypeDefs`

Type `(NamePair, NameOrigin)` is used in the following pass.  We cannot
add the origin to `NamePair` because `NamePair` is used for fields as
well.

8. `MangleNames`

The `External` AST remains the same: `DeclInfo` has a separate origin
field, and no origin is provided for `TypedefRegular` since we assume
that it is always in source.

Note that `NameOriginBuiltin` is added.