Refactor BindingSpec API
Module `HsBindgen.BindingSpec.Internal` is renamed to
`HsBindgen.BindingSpec.Private`, and the `StdLib` is moved to
`HsBindgen.BindingSpec.Private.Stdlib`.  The code is refactored so that
these modules are only used within `HsBindgen.BindingSpec` and
`HsBindgen.BindingSpec.Gen` (which must be kept separate due to the
dependency graph.

Module `HsBindgen.BindingSpec` exports a public as well as an internal
API, both using the `BindingSpec` type that wraps unresolved and
resolved binding specifications.  Functions are added to unwrap this
opaque public type and call the private implementations.