Refactor frontend trace message types
All trace message types must be in the public API so that users are able
to configure how they are handled.  Frontend pass trace message types
are refactored to be separate types, associated with the pass using a
type family instead of a data family, so that they can be re-exported
separately.  This provides an easier API for users as well as more
control over Haddock documentation.

Half of the frontend pass trace message types had constructors with
prefixes specifying the pass, while half did not.  Some types had the
same constructor names, resulting in re-export conflicts.  The types are
refactored so that all of them have constructors with prefixes
specifying the pass, which resolves the conflicts and is consistent.

The instances for the various trace classes are updated for consistency.

The trace message types are re-exported from the public API.  We should
improve how we do this.