Make macro warnings Info by default
CLI flag `--macro-warnings` restores the reparse and typecheck messages
back to warnings.

We probably want to make this flag available in the TH API as well.
This requires some refactoring, however, so it should be done in a later
commit.

Use of `withTracerCustom` (to configure a `CustomLogLevel`) requires
determining `AnsiColor`.  We should probably either refactor to avoid
this or expose `getAnsiColor`.