Rename to NonParsedDecls
`NonSelectedDecls` is renamed to `NonParseDecls`, to reflect that it
only contains information about declarations that are excluded in the
`Parse` pass.  This information is used in the `ResolveBindingSpec` pass
to resolve external bindings.  It is no longer used afterwards, so the
code in the `Select` pass to add information about non-selected
declarations is removed.