Generate stubs for global variables
On Windows, simply generating a foreign import of a global variable's address
can lead to errors (see #898). For example, given a global `int simpleGlobal`,
the following foreign import might cause an error:

-> foreign import capi safe "&simpleGlobal" simpleGlobal :: F.Ptr FC.CInt

So, instead we generate a /stub/ function that simply returns the address of
the global variable ...

-> __attribute__ ((const)) signed int *get_simpleGlobal_ptr (void) {
->   return &simpleGlobal;
-> }

... and then create a foreign import for the stub.

-> foreign import capi safe "get_simpleGlobal_ptr" simpleGlobal :: F.Ptr FC.CInt

Note that stub function also has a `const` function attribute to emphasise
that the function always returns the same address throughout the lifetime of
the program, which means we can omit the `IO` from the foreign import.

The exception to this stub generation is that we forego the stub in case the
global variable has an array type. It is not straigthforward (if possible at
all) to return pointers to arrays from a C function without losing the
array's length information.